refactor: convert useCspHook to ts
