+++ src/hooks/settingsHooks/index.js
@@ -1,4 +1,4 @@
-export { useCspHook } from "./useCspHook"
+export * from "./useCspHook"
 export { useGetSiteColorsHook } from "./useGetSiteColorsHook"
 export * from "./useUrlHook"
 export * from "./useGetSettings"
